+++ templates/BuildAction/þnameþ.action.ps1
@@ -1,0 +1,21 @@
+﻿$action = {
+	param (
+		$Parameters
+	)
+	
+	$rootPath = $Parameters.RootPath
+	$actualParameters = $Parameters.Parameters
+
+	# Insert implementation here
+}
+
+$params = @{
+	Name        = 'þnameþ'
+	Action      = $action
+	Description = 'þdescriptionþ'
+	Parameters  = @{
+		ParameterName = 'Add Description'
+	}
+}
+
+Register-PSMDBuildAction @params